Fix to close database connection for portals.
Every run of stopwatcher leaves an aborted connection in the mysql log.  This is due to it not closing the connection to the portals db at the end like it does with the main db.  I have added mydb_p.close() to the end of the script and it now functions properly.